backend/requirements.txt
fastapi[standard]
yt-dlp
uvicorn
python-multipart
ffmpeg-python
google-api-python-client
python-dotenv
sqlmodel
alembic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt==4.0.1
stripe
langchain
langchain-community
langgraph
langchain-openai
youtube-transcript-api
replicate
networkx
